refactor net utils
Refactor net utilities to a leaf package because they will create a cycle in the next commit.
